@@ -7,24 +7,22 @@ import { SlideDown } from 'react-slidedown'
 const Legend = props => 
     <div>
         <div className="d-flex">
-            <div>{props.title && <h4>{props.title}</h4>}</div>
+            <div>{props.title && <h5>{props.title}</h5>}</div>
             <div className="ml-auto">
-                <FontAwesomeIcon 
-                    onClick={() => props.toggleExtended()}
-                    className={props.extended ? 'text-primary' : 'text-muted'}
+                {props.children && <FontAwesomeIcon 
+                    onClick={e => {e.preventDefault(); props.toggleExtended()}}
+                    className={props.extended ? 'text-muted' : 'color-navigation'}
                     style={{fontSize: '1.3rem'}}
-                    icon={faInfoCircle} />
+                    icon={faInfoCircle} />}
             </div>
         </div>
-        <div className="animated fast row" style={{ overflow: 'hidden' }}>
-            <SlideDown className="animated fast">
-            {props.extended &&
-                <div className="col my-2" style={{fontSize: '.9rem'}}>
-                    {props.children}
-                </div>
-            }
-            </SlideDown>
-        </div>
+        <SlideDown className="animated fast flex-fill">
+        {props.children && props.extended && <div className="animated fast row container-fluid" >
+            <div className="col my-2">
+                {props.children}
+            </div>
+        </div>}
+        </SlideDown>
     </div>
 
 export default makeExtendable(Legend, true)
