@@ -1,3 +1,2 @@
-const animationTime = 550
-
-export { animationTime }
+export const animationTime = 550
+export const mobileWidth = 768
