@@ -1,5 +1,5 @@
 import _ from 'lodash/fp'
-
+import { get } from './DataProvider'
 /*
 * Provides basic functions to stores which need to load data from an external source
 * parameter identifier must be unique throughout the entire app
@@ -11,6 +11,7 @@ export default (identifier, getStore) => {
         switch (action.type) {
             case `${identifier.toUpperCase()}_DATA_FETCHED`:
                 return true
+            case `${identifier.toUpperCase()}_DATA_FETCH_FAILED`:
             default:
                 return state
         }
@@ -20,6 +21,9 @@ export default (identifier, getStore) => {
         switch (action.type) {
             case `${identifier.toUpperCase()}_DATA_FETCHING`:
                 return true
+            case `${identifier.toUpperCase()}_DATA_FETCH_FAILED`:
+                console.error(action.payload)
+                return false
             case `${identifier.toUpperCase()}_DATA_FETCHED`:
                 return false
             default:
@@ -27,32 +31,35 @@ export default (identifier, getStore) => {
         }
     }
 
-    const loadReducer = reducer => (state, action) => {
+    const loadReducer = (state = [], action) => {
         switch(action.type) {
             case `${identifier.toUpperCase()}_DATA_FETCHED`:
                 return action.payload
             default:
-                return reducer(state, action)
+                return state
         }
     }
 
     return {
+        getStore,
         withLoadedSelector: selectors => ({ 
             loaded: state => getStore(state).loaded,
             ...selectors
         }),
-        withLoadAction: (actions, result) => ({
+        withLoadAction: url => actions => ({
             load: () => (dispatch, getState) => {
                 if(getStore(getState()).fetching || getStore(getState()).loaded) return
-                setTimeout(() => dispatch({ type: `${identifier.toUpperCase()}_DATA_FETCHED`, payload: result}), 1000)
+                get(url)
+                    .then( result => result.json().then( data => dispatch({ type: `${identifier.toUpperCase()}_DATA_FETCHED`, payload: data })))
+                    .catch( err => dispatch({ type: `${identifier.toUpperCase()}_DATA_FETCH_FAILED`, payload: err }))
                 dispatch({ type: `${identifier.toUpperCase()}_DATA_FETCHING` })
             },
             ...actions 
         }),
-        withLoadedReducer: reducer => ({
+        withLoadedReducer: (reducer = loadReducer) => ({
             loaded,
             fetching,
-            items: loadReducer(reducer)
+            items: reducer,
         }),
         getItems: _.flow([getStore, store => store.items]),
     }
