@@ -1,5 +1,4 @@
 import React, { useEffect } from 'react'
-import _ from 'lodash/fp'
 import { connect } from 'react-redux'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCircleNotch } from '@fortawesome/free-solid-svg-icons'
@@ -8,12 +7,13 @@ const Spinner = () => <div className="text-center m-4" style={{fontSize: '1.5rem
     <FontAwesomeIcon className="fa-spin" icon={faCircleNotch} />
 </div>
 
-const needsData = (WrappedComponent, loadedSelector, loadAction) => connect(state => ({ loaded: loadedSelector(state) }), { load: loadAction })(
+const stateToProps = loadedSelector => (state, ownProps) => ({ loaded: loadedSelector(state, ownProps) })
+const needsData = (loadedSelector, loadAction, spinner = true) => WrappedComponent => connect(stateToProps(loadedSelector) , { load: loadAction })(
     ({ load, loaded, ...props }) => {
-        useEffect(() => { load() }, [load])
+        useEffect(() => { !loaded && load(props) }, [loaded, load, props])
 
-        return loaded ? <WrappedComponent {...props} /> : <Spinner />
+        return loaded ? <WrappedComponent {...props} /> : (spinner && <Spinner />)
     }
 )
 
-export default _.curryRight(needsData)
+export default needsData
