@@ -2,12 +2,13 @@ import React from 'react'
 
 export default function DashboardCard(props) {
     return (
-        <div className={`${props.color} card dashboard-card with-border`}>
+        <div className={`${props.color} card with-border overflow-hidden mt-2`}>
           <div className="card-body d-flex p-0">
-            <div className="p-3 w-100">
-                <h5 className="card-title">{props.title} {props.result && <span className="mr-3 float-right" style={{color: 'red'}}>{props.result}</span>}</h5>
-                <p style={{fontSize: '.8rem'}} className="card-text">{props.text}</p>
-                {props.children}
+            <div className={`${props.noPadding || 'p-3'} w-100`}>
+                <h5>{props.title}</h5>
+                <div>
+                  {props.children}
+                </div>
             </div>
           </div>
         </div>
