@@ -2,35 +2,56 @@ import React from 'react'
 import _ from 'lodash/fp'
 import { connect } from 'react-redux'
 import { withTranslation } from 'react-i18next'
-import SimpleBar from '../Charting/SimpleBar'
+import AnimatedDonut from '../Charting/AnimatedDonut'
 import needsData from '../Core/needsData'
-import { selectors as semesterSelectors, actions as semesterActions } from '../Exams/Semester/Store'
+import { selectors as mcSelectors, actions as mcActions } from '../Exams/MC/Store'
 import { selectors as ptmSelectors, actions as ptmActions } from '../Exams/Ptm/Store'
 import DashboardCard from './DashboardCard'
 
-const loaded = _.overEvery([ semesterSelectors.loaded, ptmSelectors.loaded ])
-const load = () => _.over([ semesterActions.load(), ptmActions.load ])
-const ptmStrongestSubject = _.flow([ ptmSelectors.getLatest, ptmSelectors.strongestSubject ])
-const stateToProps = state => ({ mcStrongestSubject: semesterSelectors.strongestSubject(state), ptmStrongestSubject: ptmStrongestSubject(state), l: semesterSelectors.getRanking(state) })
+const loaded = _.overEvery([ mcSelectors.loaded, ptmSelectors.loaded ])
+const load = () => _.over([ mcActions.load(), ptmActions.load ])
+const ptmStrongestSubject = _.flow([ ptmSelectors.getLatest, _.defaults({}), ptmSelectors.strongestSubject ])
+const stateToProps = state => ({ 
+    mcStrongestSubject: mcSelectors.strongestSubject(state), 
+    ptmStrongestSubject: ptmStrongestSubject(state),
+})
+const Label = ({ title, children }) => 
+    <div className="mt-2 w-100 d-flex flex-column">
+        <div style={{fontSize: '.8rem'}} className="text-secondary">{title}</div>
+        <div className="pt-2 d-flex align-items-center flex-grow-1" style={{fontWeight: 500}}>
+            <div>{children}</div>
+        </div>
+    </div>
+const Donut = ({ value, total }) => 
+    <div className="mt-2 w-100">
+        <div style={{height: '5rem'}}>
+            <AnimatedDonut animationTime={0} width={.22} data={[ value, total - value ]}>
+                <span style={{fontSize: '.8rem', fontWeight: 500}}>{value} / {total}</span>
+            </AnimatedDonut>
+        </div>
+    </div>
 const Strengths = _.compose([withTranslation(), needsData(loaded, load), connect(stateToProps)])(({ t, mcStrongestSubject, ptmStrongestSubject }) =>
-    <div>
-        <div className="mb-3">
-            <div style={{fontSize: '.8rem'}} className="text-secondary">{t('Stärkstes Fach in deinen gesamten MCs')}</div>
-            {mcStrongestSubject.name}
-            <SimpleBar value={mcStrongestSubject.richtig} total={mcStrongestSubject.gesamt}>{mcStrongestSubject.richtig} von {mcStrongestSubject.gesamt}</SimpleBar>
-        </div>
-        <div className="">
-            <div style={{fontSize: '.8rem'}} className="text-secondary">{t('Stärkstes Fach im letzten PTM')}</div>
-            {ptmStrongestSubject.name}
-            <SimpleBar value={ptmStrongestSubject.richtig} total={ptmStrongestSubject.gesamt}>{ptmStrongestSubject.richtig} von {ptmStrongestSubject.gesamt}</SimpleBar>
-        </div>
+    <div className="">
+        { (mcStrongestSubject || ptmStrongestSubject) && 
+        <div>
+            <div className="d-flex">
+                { mcStrongestSubject && <Label title={t('in deinen gesamten MCs')}>{mcStrongestSubject.titel}</Label> }
+                { ptmStrongestSubject && <Label title={t('im letzten PTM')}>{ptmStrongestSubject.titel}</Label> }
+            </div>
+            <div className="d-flex">
+                {mcStrongestSubject && <Donut total={mcStrongestSubject.maximalPunktzahl} value={mcStrongestSubject.ergebnisPunktzahl} /> }
+                {ptmStrongestSubject && <Donut value={ptmStrongestSubject.ergebnisRichtigPunktzahl} total={ptmStrongestSubject.maximalPunktzahl} /> }
+            </div>
+        </div>}
+        {!ptmStrongestSubject && !mcStrongestSubject && 
+            <div className="p3">{t('Es liegen noch keine Ergebnisse vor.')}</div>
+        }
     </div>
 )
 
 const Wrapper = withTranslation()(({ t }) => <DashboardCard 
     header={Math.round(Math.random() * 100) + ' p'} 
-    title={t(`Deine Stärken`)} 
-    text={t(`Dein Überblick zu deinen fächerorientierten Stärken im PTM und den Semesterprüfungen über das gesamte Studium.`)}>
+    title={t(`Starke Fächer`)} >
             <Strengths />
 </DashboardCard>)
 
