@@ -7,37 +7,35 @@ import Filter from '../../Utils/Filter'
 import Subject from './Subject'
 import { selectors, actions } from './Store'
 import css from './Subjects.module.css'
-import { minQuestions } from '../../Utils/Constants';
+import { minQuestions } from '../../Utils/Constants'
 
-const stateToProps = (state, { semester }) => _.flow(selectors.getBySemester, selectors.getSubjects, _.sortBy( s => -s.richtig / s.gesamt ), r => ({ fächer: r}))(state, semester)
-const Subjects = _.compose(needsData(selectors.loaded, actions.load), withTranslation(), connect(stateToProps))( ({ t, fächer, semester }) => {
-    const categories = _.keys(_.groupBy( s => s.kategorie, fächer ))
+const stateToProps = (state, { id }) => _.flow(selectors.getById, selectors.getSubjects, _.sortBy( s => -s.ergebnisRichtigPunktzahl / s.maximalPunktzahl ), r => ({ faecher: r}))(state, id)
+const Subjects = _.compose(needsData(selectors.loaded, actions.load), withTranslation(), connect(stateToProps))( ({ t, faecher, id }) => {
+    const categories = _.keys(_.groupBy( s => s.gruppe, faecher ))
     const [ filters, setFilters ] = useState( 
-        categories.map(c => ({label: c, pred: d => d.kategorie === c , selected: true, color: 'hsla(200, 50%, 50%, .8)', group: 'categories'}))
-        .concat({label: `mind. ${minQuestions} Fragen`, pred: s => s.gesamt >= minQuestions , selected: true, color: 'hsla(200, 50%, 50%, .8)' })
+        categories.map(c => ({label: c, pred: d => d.gruppe === c , selected: true, color: 'hsla(200, 50%, 50%, .8)', group: 'categories'}))
+        .concat({label: `mind. ${minQuestions} Fragen`, pred: s => s.maximalPunktzahl >= minQuestions , selected: true, color: 'hsla(200, 50%, 50%, .8)' })
     )
     const [ search, setSearch ] = useState('')
 
-    const filtered = fächer.filter(
+    const filtered = faecher.filter(
         _.overEvery([
             _.overSome(filters.filter( f => f.selected && f.group === 'categories').map( f => f.pred )), 
-            s => s.name.toLowerCase().indexOf(search.toLowerCase()) >= 0,
+            s => s.titel.toLowerCase().indexOf(search.toLowerCase()) >= 0,
             _.overEvery([...filters.filter( f => f.selected && f.group !== 'categories' ).map( f => f.pred ), () => true]),
         ]))
     return (
     <div className="w-100">
-        <div className="d-flex flex-column flex-lg-row">
-            <div className="">
-                <Filter filters={ filters } onUpdate={ setFilters } />
-            </div>
-            <div className="flex-grow-1 align-middle py-2">
-                <input type="text" placeholder={`${t('suchen')}...`} onKeyUp={e => setSearch(e.target.value)} className={`p-1 ${css.input}`}></input>
-            </div>
+        <div className="">
+            <Filter className="w-100" filters={ filters } onUpdate={ setFilters } />
+        </div>
+        <div className="align-middle my-2 pr-1">
+            <input type="text" placeholder={`${t('Fach suchen')}...`} onKeyUp={e => setSearch(e.target.value)} className={`p-1 ${css.input}`}></input>
         </div>
         <div className={css.subjects}>
         {filtered.map((s, rank) =>
-            <div className={`w-100`} key={s.name}>
-                <Subject extended={parseInt(rank) < 1} key={s.name} name={s.name} category={s.kategorie} rank={parseInt(rank) + 1} semester={semester} />
+            <div className={`w-100`} key={s.code}>
+                <Subject extended={parseInt(rank) < 1} key={s.code} titel={s.titel} gruppe={s.gruppe} rank={parseInt(rank) + 1} id={id} />
             </div>
         )}
         </div>
